+++ lib/glific/bigquery/bigquery_schema.ex
@@ -350,6 +350,11 @@ defmodule Glific.BigquerySchema do
         name: "contact_name",
         type: "STRING",
         mode: "REQUIRED"
+      },
+      %{
+        name: "flow_context_id",
+        type: "INTEGER",
+        mode: "NULLABLE"
       }
     ]
   end