+++ lib/glific/processor/consumer_flow.ex
@@ -54,10 +54,10 @@ defmodule Glific.Processor.ConsumerFlow do
       _ ->
         cond do
           Map.get(state, :newcontact, false) == true &&
-              Map.has_key?(state.flow_keywords, "newcontact") ->
+              Map.has_key?(state.flow_keywords["published"], "newcontact") ->
             check_flows(message, "newcontact", state, false)
 
-          Map.has_key?(state.flow_keywords, body) ->
+          Map.has_key?(state.flow_keywords["published"], body) ->
             check_flows(message, body, state, false)
 
           is_beta ->
@@ -77,7 +77,10 @@ defmodule Glific.Processor.ConsumerFlow do
 
   defp is_beta_keyword?(state, body) do
     if String.starts_with?(body, @beta_phrase) and
-         Map.has_key?(state.flow_keywords, String.replace_leading(body, @beta_phrase, "")),
+         Map.has_key?(
+           state.flow_keywords["draft"],
+           String.replace_leading(body, @beta_phrase, "")
+         ),
        do: true,
        else: false
   end
@@ -97,13 +100,15 @@ defmodule Glific.Processor.ConsumerFlow do
         {@final_phrase, body}
       end
 
-    {:ok, flow} =
-      Flows.get_cached_flow(
-        message.organization_id,
-        {:flow_keyword, body, status}
-      )
+    Flows.get_cached_flow(
+      message.organization_id,
+      {:flow_keyword, body, status}
+    )
+    |> case do
+      {:ok, flow} -> FlowContext.init_context(flow, message.contact, status)
+      {:error, _} -> nil
+    end
 
-    FlowContext.init_context(flow, message.contact, status)
     {message, state}
   end