+++ test/glific/flows/message_vars_parser_test.exs
@@ -6,12 +6,13 @@ defmodule Glific.Flows.MessageVarParserTest do
 
   test "parse/2 will parse the string with variable", attrs do
     # binding with 1 dots will replace the variable
-    parsed_test =
-      MessageVarParser.parse("hello @contact.name", %{"contact" => %{"name" => "Glific"}})
+    assert "hello Glific" ==
+             MessageVarParser.parse("hello @contact.name", %{"contact" => %{"name" => "Glific"}})
 
-    MessageVarParser.parse("hello @organization.name", %{"organization" => %{"name" => "Glific"}})
-
-    assert parsed_test == "hello Glific"
+    assert "hello Glific" ==
+             MessageVarParser.parse("hello @organization.name", %{
+               "organization" => %{"name" => "Glific"}
+             })
 
     # binding with 2 dots will replace the variable
     parsed_test =
@@ -70,5 +71,7 @@ defmodule Glific.Flows.MessageVarParserTest do
     ## for contact groups
     conatct_fields = Contacts.get_contact_field_map(contact.id)
     assert MessageVarParser.parse("@contact.in_groups", %{"contact" => conatct_fields}) == "[]"
+    assert MessageVarParser.parse("Hello world", nil) == "Hello world"
+    assert MessageVarParser.parse("Hello world", %{}) == "Hello world"
   end
 end